fixing some windows stuff with the debugger
